+++ changelog.md
@@ -5,10 +5,13 @@ This project follows semVer, where:
 - **[patch]** is a bugfix
 
 <!-- [unreleased] 
+-->
+
 ### v6.12.0
+- **[new]** - support period-seperated short-iso format
 - **[new]** - support fortnight in add/remove methods
 - **[new]** - support 'tues' 'thurs' in day() method
--->
+- **[new]** - support '2002-06' truncated iso input
 
 ### v6.11.0
 - **[change]** - support 6-digit millisecond, and lowercase iso